Remove unsubstituted_url_variables handler
This was added by https://bugzilla.mozilla.org/show_bug.cgi?id=1282569
because we were apparently encountering 400s tens of thousands of
times per day. However, it isn't clear why. Removing the handler falls
back to the real one, which fails to find any information and
generates the exact same response. This adds a tiny amount of database
load but simplifies the code.