Simplify rowsToDicts
Converting rows to dicts one field at a time is somewhat wasteful. We
can just run the dict() builtin.

There's no need for this to be a decorator -- it doesn't change any
behavior of the function except its return value. Besides that, it's
only used on one function, which only really has one return statement,
so just use it there.