Remove no-ops
In both places it occurs, this piece of code converts a list of dicts
to a list of the exact same dicts.

Perhaps it would have been valuable if it had applied a whitelist or
if the original objects had not also been dicts, but they come from
getOrderedRules, which calls select, which calls rows_to_dicts.