Update to Node 4.x
Move off of a no-longer-supported version of NodeJS to a version that
is still at least barely supported.

Debian's jessie-backports repo has an updated version of NodeJS, but
it wasn't possible to use npm with it because of
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=861178, which
prevents the installation of nodejs-dev, which prevents the
installation of node-gyp, on which npm depends. As a result, we
introduce a Nodesource repository, which seems to work fine.

This means NodeJS is no longer the limiting factor for ES6
support. Now PhantomJS is.