Remove whitelists (#401). r=bhearsum,allan-silva
* Simplify rowsToDicts

Converting rows to dicts one field at a time is somewhat wasteful. We
can just run the dict() builtin.

There's no need for this to be a decorator -- it doesn't change any
behavior of the function except its return value. Besides that, it's
only used on one function, which only really has one return statement,
so just use it there.

* Remove no-ops

In both places it occurs, this piece of code converts a list of dicts
to a list of the exact same dicts.

Perhaps it would have been valuable if it had applied a whitelist or
if the original objects had not also been dicts, but they come from
getOrderedRules, which calls select, which calls rows_to_dicts.

* Turn whitelist into a blacklist

These whitelists appear to actually have a function, which is namely
to not serialize the data field from a release. I believe this is
being done for efficiency's sake, since data is the largest field in
a record. Rewrite this as a blacklist to make it clear.