Add signoff requirements to admin /releases view
Rules had to be computed on the client side, because editing a rule
can mean moving it from one channel to another, which means two sets
of signoff requirements. However, the set of rules which refer to a
specific release cannot be edited by the front-end, so a release is
always affected by the same rules, which means it will always have the
same signoff requirements. Because they are relatively static, we can
compute them on the server side and save ourselves some code.

Because this does a bunch of DB calls in a loop, it may be too
expensive to calculate in this way, but let's start here.