add GlobVersion to support version globs.
For the System Addons in Balrog work, we currently find ourselves
needing to specify a `<next_version` rule for every Firefox version,
just to limit the rules to a single Firefox major version, even if many
of the rules are noop.

We can target our rules more selectively if we're able to use a
`version: 84.*` rather than rely on having both a `version: <84` and a
`version: <85` rule.