Support globbed version matching.
In SystemAddons, we want to target specific version ranges, e.g. `80.*`.
Previous to this patch, that meant we needed both a `<80` rule to
prevent sending the same System Addon(s) to previous versions, and a
`<81` rule to target the `80.*` range. This worked while we maintained
SystemAddons rules manually, but will make it difficult to rewrite rules
automatically.

This patch adds support for `version: 80.*` or `version: 80.1.*` type
globs.