Switch CI to taskgraph
Use taskgraph instead of defining tasks directly in .taskcluster.yml

This requires duplicating the Dockerfiles for backend and agent under
taskcluster/docker/ because of the different expectations for things like
paths.

The push-dockerimage.sh scripts are replaced by a skopeo-based script stolen
from shipit, which uses the image built in a separate task and just overwrites
its version.json, instead of building a completely new image before pushing it
to docker hub.

The balrog-specific taskgraph code is necessary to:
- add a "pull_request_number" decision task parameter and pass it to
  coveralls via environment
- add a "release" target_tasks method for github-release events, and
  filter tasks on github-release based on the git tag name (run-on-releases,
  similar to the existing run-on-git-branches)
- pass some params into the push-image script's environment