tests: add a check that we can deal with encoded colon and comma in system capabilities
Bug 1620327 wants to encode the update URL properly, so let's make sure it works.

As far as I can tell this doesn't require changes to the application
code, because the CGI and wsgi specs say PATH_INFO is not URL-encoded,
so we don't have to deal with this at all.