Remove assumptions about release blobs from rule evaluation (#3065)
`evaluateRules` can't reliably call `blob.getApplicationVersion`, because that
method is specific to certain types of blobs.  Instead, we change the
`blob.shouldServeUpdate` method to return a tristate instead of a boolean: `No`
means no update should be served.  `Maybe` means the update can be served but
the pin version should be checked.  `Yes` means the update should be served.

In the `Maybe` case, `evaluateRules` can look for a release corresponding to
the pin version and return that instead if found.

Closes #2985