fix: guard against unparsable nightly release names in cleanup script (#3086)
I discovered this issue while testing the cleanup script in stage, which had a release with a name like "Firefox-mozilla-central-nightly-20170731100325-2". The dry run of the cleanup script worked fine, but when it tried to delete it threw an error about not being able to parse it as a date. After some digging, I discovered that this is because the dryrun's `SELECT` uses an index, while the `DELETE` does not. (I don't know why this is - but it doesn't affect the behaviour of the script beyond making the `DELETE` look at some rows that it won't be deleting anayways...)

In any case, we may as well make sure the names we're about to parse are actually parsable.