Add support for mirror URLs for GMP plugins.
This patch adds support for alternative URLs from which the GMP plugin
can be downloaded. This is useful for GMP plugins, particularly
Widevine, because there are many different mirrors from Google. Some
users have trouble accessing the preferred mirror due to networking
issues, so it would be ideal to supply alternative domains that are more
likely to resolve/be reachable as a backup.