chore: update dependabot to use groups
Currently we're being spammed with PRs. This should get dependabot to
batch PRs together such that all minor/patch updates are done in a
single PR. Major updates should still be individual.

It's not clear to me whether dependabot truly supports
`pip-compile-multi`, so we may not be able to start accepting this PRs
until we use something more standard.

At least this will reduce the number of possible PRs from 150 -> 15.