Switch from react-auth0-components to auth0/auth0-react
The former has been unmaintained for 5 years. I was looking at why the
login button wasn't working on first load and fell into a rabbit hole...

Regarding the update, this is fairly "straightforward". We now use the
context provided with the auth0-react library instead of reinventing the
wheel and having our own. The auth0 library also sets up its own
callback things so we don't need to set that up manually anymore either.

The main pain point was figuring out how to know within the axios
interceptors whether we were logged in or not as calling
`getAccessTokenSilently` will try to log you in if you're not (do not
ask me why... I already didn't have hair anymore before but it's way
worse now). Turned out that we could use `getIdTokenClaims` for this. I
would've loved to set/unset the interceptor based on the `isAuthenticated`
boolean we can get from auth0 but the interceptor also sets the base URL
so we can't do that or it breaks the heartbeat request.