Don't use the `db_schema` fixture in `guardian_db` (#3448)
The db_schema fixture is session scopes while the guardian_db one is
module scoped. If the guardian_db fixture is used in tests that get
run out of order, the module fixture gets called again and explodes
because it's inserting into the session database again.

For example if tests are ran like this:

- test_json:A
- another_module:Foo
- test_json:B

It would result in:

- db_schema
- guardian_db
- test_json:A
- another_module:Foo
- guardian_db # BOOM
- test_json:B

This should hopefully fix the intermittent we've been seeing in test_json