fix: reduce the number of requests needed to populate the the Users UI view (#3440)
* fix: make getAllPermissions consistent with getUserPermissions

It looks like this is currently unused, but will be useful for the subsequent change of retrieving all user permissions at once. It needs to:
* Do a permissions check to make sure the retrieving user is allowed to view other users' permissions
* Return the same data in the same format as getUserPermissions

* feat: return user permissions as part of `/users` endpoint

This will allow us to avoid making one request per user in the `/users` view of the UI, which will hopefully allow us to fix https://github.com/mozilla-releng/balrog/issues/1043.

* feat: fetch data needed for users view in 2 requests