refactor: make get_update_blob more readable by moving some inner logic to helpers
Note the slight and subtle behaviour change here of not updating `eval_metadata` in the `response_products` block. This powers the `Rule-ID` and `Rule-Data-Version` headers we set, which are already incomplete for this case (we'll use the versions from the last product processed). With this change, we'll use the versions from the original query (eg: of the `GMP` product). Given that this already didn't work properly for this case, I'm inclined to not worry about this.