feat: add request timing and response code stats for the public app (#3459)
* feat: run statsd and graphite in docker compose

This allows us to better validate changes to metrics we're publishing.

* refactor: make get_update_blob more readable by moving some inner logic to helpers

Note the slight and subtle behaviour change here of not updating `eval_metadata` in the `response_products` block. This powers the `Rule-ID` and `Rule-Data-Version` headers we set, which are already incomplete for this case (we'll use the versions from the last product processed). With this change, we'll use the versions from the original query (eg: of the `GMP` product). Given that this already didn't work properly for this case, I'm inclined to not worry about this.

* feat: add timers for public balrog app response processing

The changes look big, but really it's just wrapping existing code in context managers.

* feat: log response codes to statsd

To start with, just by the first segment of the path. We can get more granular in the future if we find a use case for it.