feat: add statsd timings for admin endpoints
The most important things we need to time here are the /releases endpoints, as they are (mostly) the only ones used frequently, and we also believe them to be some of the least performant ones. I don't think there's much downside to timing just about everything though, especially considering most of them are rarely used, and won't generate data most of the time.

One way I looked at doing this and ultimately rejected was using a `timer` decorator on specific handlers. This had the upside of allowing custom metric names to be set as an argument (which would've avoided renaming everything), but I decided that decorating everything was both ugly, and not very future proof (we'd probably forget to decorate things in the future).