Move the frontend configuration directly into the task environment
This was set by fetching a secret from taskcluster containing only
public values, then put into a .env file which was finally exposed
through `dotenv` in `packages.json`. Since all of those values are
actually public, there's no need to go through that whole process and we
can store those values into the task definition.

This brings deployment of balrog more in line with what shipit does
(although shipit stores those in the taskgraph config and then copies it
into the task environment but I don't see a point in doing that). It
also means that we can now search for those values... When the dotenv PR
broke staging, it took a while before someone figured out that those
values were coming from a secret because there were no mentions of them
anywhere that was deployment related. And it'll allow us to remove the
dependency of dotenv.

I decided to put `GCS_NIGHTLY_HISTORY_BUCKET` and
`GCS_RELEASES_HISTORY_BUCKET` in the prod config despite them being the
default in the webback config so I could test for their existence in the
deploy script.