Replace eslint/prettier with biome
When I looked at updating eslint, it got... Very involved and I ended up
on the eslint-config-airbnb package which contained most of the used
rules which didn't even support the latest version of eslint. Upon
searching, I found biome which seems to be the goto linter these days in
the js ecosystem (gotta change all your tools every 3 years you know,
keep it on your toes). Since it's way faster, seems to be stricter,
doesn't pull in half of npmjs, since I needed to rewrite the entire
eslint config to begin with and since we didn't care about it to begin
with and only used it because neutrino was, this seemed like a good time
to switch over. The configuration is much much much simpler (as
evidenced by the diff between biome.jsonc and .eslintrc.js, and that's
not even counting the dependency hell that eslint had just for rules...)

Since biome also replaces prettier, that's one less dependency too,
which I'm always happy with.

Now for the changes themselves. I left all recommended rules enabled
except for 2.

- `useExhaustiveDependencies` for react because I'm scared. It gives us
  84 errors about react hooks that do not have the proper dependencies.
  Those might actually be hiding bugs where components might not
  rerender when state changes. However, those issues aren't new and I
  don't feel like breaking everything right now, I just disabled the
  rule entirely and added a FIXME to be clear that it should be enabled

- `noUselessFragments` for react because I think our react version is
  too old? All the ones I've tried to remove ended up in syntax errors
  and I know that stuff changed around jsx in later react versions.

I let biome fix the rest which seems harmless enough although it makes
for a bit of a noisy diff. I didn't bother with formatting those fixes
as I intend on making a second commit to run the autoformatter because
there's no way I'm getting to the exact same thing as prettier anyway...

We do lose one "functionality" from eslint/prettier which is that it used
to emit warnings on webpacking. I personally found that very annoying
anyway as it was hiding real errors but that means that those will most
likely be caught by CI now unless people think about running `yarn
lint/yarn format` locally.