Run formatter
I made most of it conform with prettier to reduce the diff size but I
didn't bother much more. I'm of the opinion that formatting doesn't
really matter as long as it's consistent and if it feels like every
function needs to use `(param) =>` instead of `param =>` then I'm not
going to fight on it. If someone has strong opinions about this, feel
free to change the config.