Remove mozilla frontend infra components (#3499)
* Replace `Label` from mozilla-frontend-infra/components

We need to get rid of mozilla-frontend-infra/components, this is the
first step. I basically imported the component and stripped the unused
parts. This doesn't change anything rendering wise

* Switch from mozilla-frontend-infra's `MuiErrorPanel` to raw `Alert`

* Replace all mozilla-frontend-infra's Spinner with `CircularProgress`

* Remove the dependency on mozilla-frontend-infra

We depend on mdi-react and it used to work because it was pulled through
mozilla-frontend-infra/components so I just added it as an explicit
dependency now