Replace unidiff usage with diff and update the latter (#3511)
* Replace `unidiff` usage with `diff`

We already depend on `diff for the releases diffs. This was just a
second dependency used to do the same thing, and unidiff depends on diff
(an outdated version at that), so just remove it.

I ran into react-diff-view not liking createPatch adding `Index:
file\n=============================` at the top even though that seems
to be valid in a unified diff (patch(1) ate it without complaining) so I
just those two off. If those lines ever come to disappear then we'll eat
the entire header instead which will make react-diff-view very angry so
I'm not worried about eating two lines off a diff by mistake.

The other difference is that when the files match (oldText, newText are
empty strings), unidiff used to return an empty string but createPatch
returns a header which makes parseDiff throw some exception, so we
catch this and return an empty string which is what it did before

* Update diff