Replace downshift with material UI's Autocomplete (#3524)
I'm assuming that Autocomplete didn't exist at the time this was written
because it basically does what the wrapper was doing but better.
Downshift was severely out of date, but one less dependency makes me
happier than one less outdated dependency.

This actually has better usability on multiselects because you can use
the dropdown list to deselect things, it'll also highlight the already
selected ones. It also fixes fields that should have been autofocused
according to the code (like when creating a release, the product name is
marked as autofocus, but that didn't work prior to this, it might be
something to keep in mind as we might not actually want said autofocus
if no one's complained before).