Replace mdi-react usage with mui/icons-material
This is consistent with shipit. We had both dependencies but only need
one. He we didn't have usage of mui/icons-material but it's the one
provided by material UI so I think we should be using it instead of some
third party doing the same thing given the icons set is the same and
is coming from the same place.

As far as I can tell, this commit doesn't change *anything* display wise
as we were using a fairly up to date mdi-react already.