Ignore errors from `getAccessTokenSilently` (#3526)
We cannot do anything about those and they will log the user out which
will prevent the call from failing again (as it'll reset the claim
token). If we don't ignore it, it logs the user out anyway but also
shows a "Are you sure you're connected to the VPN" error.

This is way easier than what I did in shipit because all API requests in
balrog are authenticated including the `__heartbeat__` so we can just
use the fact that the heartbeat tries to get a token before actually
calling the route to check for the token on load