Switch dependency management to uv (#3529)
The python 3.9 requirements on the workspace stems from balrogclient. Since a workspace can only declare one required python version and has to be the minimal one. However that only influences what the lockfile contains and the actual python version used is decided by the user (or in our case the version of python in our containers).
We can probably bump that version to 3.11 but I'd rather do that in a separate PR.