Pass the state correctly to the `Link` component
This fixes a regression from c3f865c7229f128fb99de73e66891f76d918ff4e
where updating react-router silently broke keeping some state across
navigation. The issue is that react router's API changed the state
property from being nested within the `to` object to being directly on
the routing components and the `to` component becoming a string. We
didn't see this at all because of the `Link` wrapper that was doing some
massaging on the `to` property when it wasn't a string, so we were just
silently discarding the state parameter.

I removed said massaging to since we're passing strings to `to`
everywhere now as we should.

Fixes #3547