feat: add support for caching in Redis
This commit makes it possible to cache and retrieve objects from Redis, while maintaining the benefits of a local in-memory cache. It is off in production code for the moment, but the tests have been updated to use it, to make sure it stays well tested until we swap production over. (I expect the period of time that both are technically supported to be quite short, so I didn't bother adding separate tests for it; if this turns out to be untrue, this should be revisited.)