feat: add RedisCache class
This provides a compatible interface with the ExpiringLRUCache that we currently use, making it easy to turn on. It is off in production code for the moment, but the tests have been updated to use redis, to make sure it stays well tested until we swap production over. (I expect the period of time that both are technically supported to be quite short, so I didn't bother adding separate tests for it; if this turns out to be untrue, this should be revisited.)