feat: add support for using redis for public app instead of in memory caches (#3556)
* feat: add redis container

* feat: add support for caching in Redis

This commit makes it possible to cache and retrieve objects from Redis, while maintaining the benefits of a local in-memory cache. It is off in production code for the moment, but the tests have been updated to use it, to make sure it stays well tested until we swap production over. (I expect the period of time that both are technically supported to be quite short, so I didn't bother adding separate tests for it; if this turns out to be untrue, this should be revisited.)

* feat: make it (optionally) possible to use redis for the public app

Controlling this with an environment serves two purposes:
1) Lets us turn it on for stage without turning it on for prod
2) Allows us to turn it off with just an infra change rather than a code backout