fix: make blobs safely unpicklable
When we attempted to turn on redis caching in production today an interesting new issue came up where we got tracebacks such as:
```
Traceback (most recent call last):
  File "flask/app.py", line 1823, in full_dispatch_request
    rv = self.dispatch_request()
  File "flask/app.py", line 1799, in dispatch_request
    return self.ensure_sync(self.view_functions[rule.endpoint])(**view_args)
  File "connexion/decorators/decorator.py", line 68, in wrapper
    response = function(request)
  File "connexion/decorators/uri_parsing.py", line 149, in wrapper
    response = function(request)
  File "connexion/decorators/validation.py", line 399, in wrapper
    return function(request)
  File "connexion/decorators/produces.py", line 41, in wrapper
    response = function(request)
  File "connexion/decorators/parameter.py", line 120, in wrapper
    return function(**kwargs)
  File "auslib/web/public/helpers.py", line 17, in wrapper
    return f(*args, transaction=transaction, **kwargs)
  File "auslib/web/public/client.py", line 155, in get_update_blob
    release, update_type, eval_metadata = AUS.evaluateRules(query, transaction=transaction)
  File "auslib/AUS.py", line 152, in evaluateRules
    candidate = blob.shouldServeUpdate(updateQuery)
  File "auslib/blobs/apprelease.py", line 222, in shouldServeUpdate
    self.log.debug("Matching rule has no application version, will not serve update.")
AttributeError: 'ReleaseBlobV1' object has no attribute 'log'
```

The `log` attribute on blobs is set as a class attribute during `__init__` (https://github.com/mozilla-releng/balrog/blob/9c916ab4666eeb57d38ece395112f613bae87e77/src/auslib/blobs/base.py#L156). According to the comment above that, and confirmed with some digging, this was done because at some earlier point in time we required these objects to be picklable due to the `deepcopy` that happens when caching in the admin app (see https://github.com/mozilla-releng/balrog/pull/46), and storing the `log` as an instance level attribute prevented that from working.

The impact of this is that `log` may not be set when a blob is unpickled in a new process unless the class has been initialized by some other request in the process. In production, we seemed to get a very small number of these errors, presumably as processes cycled.

As of Python 3.7, loggers are picklable (https://github.com/python/cpython/issues/74705), and we can now use `__init_subclass__` to achieve this (which is run when something in unpickled). A test has been added that simulates the presumed production scenario as best as we can.