Batch required signoffs for releases
This is the same idea as for the product signoffs. Instead of doing one
query per release, we pre-fetch everything in one query then build a
mapping. To avoid duping the logic between get_releases and
Rules.getPotentialRequiredSignoffs, I extracted it into a new helper
(I'm not expecting that logic to ever change but you never know...)

With 60k releases/10k rules, this gets `get_releases` from 167s to
0.21s. And no that is not a typo.

Fixes #2995