Reduce the memory impact of compression in extract-active-data (#3586)
We're running this as a cron and it's been OOM'ing a lot. Doubling the
memory usage on that job wasn't even enough so let's just accept the
slowdown (it's a daily cronjob, as long as it finishes within 24h, it's
fine) and make it use only one core instead of however many cores it
sees as available.