Remove our custom remToPx to make rendering virtual lists faster
While profiling the page rendering for rules and releases, this was
appearing as taking ~80% of the time. Caching the font size was enough
to make it go away and make the pages usable even with 60k items,
however don't even need that since MUI exposes that value in
theme.typography.htmlFontSize already, so let's just use that.