Reduce the overscan used for `VariableSizeList` (#3590)
In ac4ea53379045a73cacfe81f264f2eb5a7d7b474 I changed it from 5 to 20
because I think I had issues with the element sizing during dev and I
had figured that the overscan might have been the culprit. And because
everything was working, I kept it there. Fast forward to me testing the
rules page with 10k items... Reducing the overscan from 20 to 5 makes
the page go from sluggish to scroll to smooth. Turns out rendering the
rule cards is expensive. It's definitely not necessary to have such a
large overscan so I see no drawbacks to doing this.