Add a way to see the diff for the currently edited rule
One thing that bothered me recently while editing rules is that the form
is very unhelpful wrt what got modified already. There's no way to know
if you misclicked on a `x` and no way to see the final state of the rule
before you save it. This adds a button at the bottom showing you the
number of changes and when clicked, pops a RuleCard comparing the
current state of the rule with the one pre-edition.

Worth noting that when editing a rule with scheduled changes, it will
compare it against the currently used one instead of the one containing
the scheduled changes.

The one thing that is a bit unfortunate is that I had to rewrite how the
fabs work on this page to avoid having a third CSS rule to align a
button... That's because the SpeedDial really wants its place in the
bottom right and if we don't  force the fab container 10rem to the right
when it's present then it will be displayed below the speed dial.